+++ example/android/app/src/main/kotlin/team/moroz/headphones_detection_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package team.moroz.headphones_detection_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()